# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kaas-ui",
    "description": "A read-only, multi-cluster Kafka UI. Every data route is a GET; there is no mutating endpoint in this document because there is none in the router.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/environments/{env}/clusters/{id}/brokers": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "`GET /api/clusters/{id}/brokers`",
        "operationId": "brokers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Brokers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Broker"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/environments/{env}/schema-registries/{registry}/subjects": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "`GET /api/environments/{env}/schema-registries/{registry}/subjects`",
        "operationId": "list",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registry",
            "in": "path",
            "description": "Schema registry id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match on the subject",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "asc | desc, by name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "details",
            "in": "query",
            "description": "Fetch id, format, version and compatibility",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subjects of the registry this cluster uses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectList"
                }
              }
            }
          }
        }
      }
    },
    "/api/environments/{env}/schema-registries/{registry}/subjects/{subject}/versions": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "`GET /api/environments/{env}/schema-registries/{registry}/subjects/{subject}/versions`",
        "operationId": "versions",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registry",
            "in": "path",
            "description": "Schema registry id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subject",
            "in": "path",
            "description": "Subject name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every registered version of a subject",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_Broker": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A broker, as the snapshot reports it — optionally enriched.",
              "required": [
                "nodeId",
                "host",
                "port",
                "isController",
                "leaderPartitionCount",
                "replicaPartitionCount"
              ],
              "properties": {
                "host": {
                  "type": "string",
                  "description": "Advertised host."
                },
                "isController": {
                  "type": "boolean",
                  "description": "Whether this broker is the controller."
                },
                "isFenced": {
                  "type": [
                    "boolean",
                    "null"
                  ],
                  "description": "Whether the controller has fenced it.\n\n`None` on a cluster that does not implement `DescribeCluster`: absent,\nnot false. Rendering an unknown as \"healthy\" is the one thing a fleet\ndashboard must never do."
                },
                "leaderPartitionCount": {
                  "type": "integer",
                  "description": "Partitions this broker leads.",
                  "minimum": 0
                },
                "nodeId": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Node id."
                },
                "port": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Advertised port."
                },
                "rack": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Rack, when the broker declares one."
                },
                "replicaPartitionCount": {
                  "type": "integer",
                  "description": "Partition replicas hosted here.",
                  "minimum": 0
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
      "ErrorKind": {
        "type": "string",
        "description": "The error taxonomy as the frontend sees it.\n\nA projection of [`kafka_conn::Error`]'s variants, not a re-listing: the\nfrontend renders a transport failure and a closed connection identically,\nso they share a kind.",
        "enum": [
          "transport",
          "timeout",
          "auth",
          "authorization",
          "broker",
          "decode",
          "unsupported",
          "invalid",
          "readOnly",
          "other"
        ]
      },
...
      "RegistryCard": {
        "type": "object",
        "description": "One schema registry, as the browser and the fleet render it.\n\nA view of a **registry**, reached through a cluster. Two clusters that\nreference `dev` show the same subjects, so the card says which registry is\nanswering rather than implying the subjects belong to the cluster whose nav\nyou arrived through.",
        "required": [
          "id",
          "name",
          "url",
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "What went wrong, where something did."
          },
          "id": {
            "type": "string",
            "description": "The configured id, which is what a cluster's `schema_registry:` names."
          },
          "name": {
            "type": "string",
            "description": "The name to render."
          },
          "status": {
            "$ref": "#/components/schemas/RegistryStatus",
            "description": "Whether it has been reached, and how it failed if it has not."
          },
          "url": {
            "type": "string",
            "description": "Where it is. Shown because a registry that is answering the wrong API\nis diagnosed by looking at its url."
          }
        }
      },
      "RegistryStatus": {
        "type": "string",
        "description": "The registry health state, flattened for the wire.",
        "enum": [
          "unprobed",
          "ready",
          "unreachable",
          "misconfigured"
        ]
      },
...
      "ResourceError": {
        "type": "object",
        "description": "One resource's failure, named.",
        "required": [
          "resource",
          "kind",
          "message",
          "retriable"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "The broker's error code name, when this build has one for it."
          },
          "codeNumber": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The error code's number.\n\nPresent even when `code` is null: against a broker newer than the codec\n`ErrorCode::Unknown(i16)` is all there is, and the number is the only\nsearchable thing."
          },
          "kind": {
            "$ref": "#/components/schemas/ErrorKind",
            "description": "The taxonomy."
          },
          "message": {
            "type": "string",
            "description": "The error, rendered."
          },
          "resource": {
            "type": "string",
            "description": "Which resource — a topic name, a broker id, a group id."
          },
          "retriable": {
            "type": "boolean",
            "description": "Whether retrying this one resource is worth offering."
          },
          "unsupportedApi": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnsupportedApiDetail",
                "description": "Both version ranges, when the failure was an unsupported api."
              }
            ]
          }
        }
      },
...
      "SchemaFormat": {
        "type": "string",
        "description": "Which of the three registry formats a schema is in.",
        "enum": [
          "avro",
          "protobuf",
          "json"
        ]
      },
...
      "SchemaReference": {
        "type": "object",
        "description": "A schema this one refers to, stored separately by the registry.\n\nA resolver that fetches only the id in the payload decodes the simple\ntopics and fails on the interesting ones — so these are shown, and they are\nfollowed when decoding.",
        "required": [
          "name",
          "subject",
          "version"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name the referring schema uses — a fully-qualified Avro name, or a\n`.proto` import path."
          },
          "subject": {
            "type": "string",
            "description": "The subject the referenced schema is registered against."
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "description": "Its version.",
            "minimum": 0
          }
        }
      },
...
      "SubjectDetail": {
        "type": "object",
        "description": "The versions of one subject, newest last.",
        "required": [
          "subject",
          "versions",
          "errors"
        ],
        "properties": {
          "compatibility": {
            "type": [
              "string",
              "null"
            ],
            "description": "The compatibility mode, where the registry reports one."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "Versions the registry listed and would not hand over, with why.\n\nPartial failure is a result: a subject with thirty versions of which\none is unreadable renders twenty-nine and names the thirtieth."
          },
          "registry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RegistryCard",
                "description": "The registry answering."
              }
            ]
          },
          "subject": {
            "type": "string",
            "description": "The subject that was asked for."
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectSchema"
            },
            "description": "Every registered version, oldest first, with its text."
          }
        }
      },
      "SubjectList": {
        "type": "object",
        "description": "The subjects one registry holds, and which registry that is.",
        "required": [
          "subjects",
          "total"
        ],
        "properties": {
          "compatibility": {
            "type": [
              "string",
              "null"
            ],
            "description": "The registry-wide compatibility mode, when `details` was asked for.\n\nWhat every row without an override of its own is governed by, so the\ntable can say `BACKWARD` on a registry where nobody has set a\nper-subject rule instead of a column of blanks."
          },
          "registry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RegistryCard",
                "description": "The registry answering, or `None` where this cluster references none."
              }
            ]
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectRow"
            },
            "description": "The page of subjects, filtered and ordered as asked."
          },
          "total": {
            "type": "integer",
            "description": "How many subjects matched before paging.",
            "minimum": 0
          }
        }
      },
      "SubjectRow": {
        "type": "object",
        "description": "One row of the subject table.\n\nEverything but the name is `Option`, and for one reason: the name is in the\nlisting the registry already gave us, and the rest is a call per subject.\nThe client asks for the page first and the columns second.",
        "required": [
          "subject",
          "compatibilityInherited"
        ],
        "properties": {
          "compatibility": {
            "type": [
              "string",
              "null"
            ],
            "description": "The compatibility mode governing this subject."
          },
          "compatibilityInherited": {
            "type": "boolean",
            "description": "Whether that mode is the registry's rather than this subject's own.\n\nTwo subjects reading `BACKWARD` are not the same fact if one of them\nset it and the other is inheriting it — the second changes when the\nregistry default does."
          },
          "format": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaFormat",
                "description": "Which of the three formats the newest version is."
              }
            ]
          },
          "id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The global schema id of the newest version — the number the wire\nformat carries.",
            "minimum": 0
          },
          "subject": {
            "type": "string",
            "description": "The subject name."
          },
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The newest version number.",
            "minimum": 0
          }
        }
      },
      "SubjectSchema": {
        "type": "object",
        "description": "One registered version of a subject.",
        "required": [
          "subject",
          "version",
          "id",
          "format",
          "schema",
          "references"
        ],
        "properties": {
          "format": {
            "$ref": "#/components/schemas/SchemaFormat",
            "description": "Which of the three formats."
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "The global schema id — the number the wire format carries.",
            "minimum": 0
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchemaReference"
            },
            "description": "Subjects this one refers to."
          },
          "schema": {
            "type": "string",
            "description": "The schema itself, as the registry stores it. Text rather than parsed,\nbecause this is what Monaco highlights and what a diff is taken over."
          },
          "subject": {
            "type": "string",
            "description": "The subject it is registered against."
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "description": "The version within that subject.",
            "minimum": 0
          }
        }
      },
...
      "UnsupportedApiDetail": {
        "type": "object",
        "description": "The two version ranges an [`Error::UnsupportedApi`] carries.\n\nThe pair is a diagnosis rather than a failure, and the three cases render\ndifferently:\n\n* `ours: None` — this build has no schema for the key; bump the codec.\n* `broker: None` — the cluster does not implement it at all.\n* both present but disjoint — the cluster is behind.",
        "required": [
          "api",
          "apiKey"
        ],
        "properties": {
          "api": {
            "type": "string",
            "description": "The api key's name."
          },
          "apiKey": {
            "type": "integer",
            "format": "int32",
            "description": "Its number, which is the searchable thing when the name is unknown."
          },
          "broker": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "What the broker advertises, as `[min, max]`."
          },
          "ours": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "What this build speaks, as `[min, max]`."
          }
        }
      }
    }
  },
...
}